% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.eggSim}
\alias{autoplot.eggSim}
\title{Title}
\usage{
\method{autoplot}{eggSim}(object, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
means <- eggSim(c(0.2,0.1,0.05), cv_reduction=0, R=10^2, parallelise=FALSE)
ggplot2::autoplot(means)

}
