% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_result.R
\name{classify_result}
\alias{classify_result}
\title{Extract clasification and confidence intervals, using the same underlying code as for survey_sim}
\usage{
classify_result(
  pre,
  post,
  efficacy_expected = 0.95,
  efficacy_lower_target = 0.85,
  alpha = 0.025
)
}
\arguments{
\item{pre}{integer vector of pre-treatment data}

\item{post}{integer vector of post-treatment data}

\item{efficacy_expected}{desired expected efficacy target (proportion, not \%)}

\item{efficacy_lower_target}{desired lower efficacy target (proportion, not \%)}

\item{alpha}{desired confidence level (0.025 = 95\% CI)}
}
\description{
Extract clasification and confidence intervals, using the same underlying code as for survey_sim
}
