% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_k.R
\name{combined_k}
\alias{combined_k}
\title{Calculate the effective overall k based on coefficients of variation}
\usage{
combined_k(cv_between, cv_within, cv_slide, cv_reduction)
}
\arguments{
\item{cv_between}{A vector of CV reflecting variation in EPG between individuals in each community}

\item{cv_within}{Day-to-day variation in EPG within an individual}

\item{cv_slide}{Variation between faecal samples from the same individual and day}

\item{cv_reduction}{Variation in efficacy between individuals}
}
\value{
a vector of pre_k and post_k
}
\description{
Calculate the effective overall k based on coefficients of variation
}
\examples{
ERRparameters(100, 10, 2, 2.5, 0.4, TRUE)

}
