% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_means.R
\name{design_means}
\alias{design_means}
\title{Calculate the mean egg counts/rates according to various design types}
\usage{
design_means(
  simdata,
  design = c("NS", "SS", "SSR1", "SSR2", "SSR3"),
  budget = 600,
  second_slide_cost = 0.621,
  max_screen = 0.9,
  count = TRUE,
  log_constant = if (count) 1 else 0,
  screen_threshold = 0
)
}
\arguments{
\item{simdata}{The data as simulated by cgpDataSim or clpDataSim}

\item{count}{Logical flag to base means on count data or the underlying rates}

\item{log_constant}{A constant to add to the count data before calculating geometric means (ignored if count==FALSE)}

\item{screen_threshold}{The threshold count on which to screen individuals}

\item{N}{A named vector of number of included individuals depending on design type}
}
\value{
A data frame of summary statistics
}
\description{
Calculate the mean egg counts/rates according to various design types
}
\examples{

}
