% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggSim-package.R
\docType{package}
\name{eggSim-package}
\alias{eggSim-package}
\title{eggSim: Simulating Datasets Representing Egg Reduction Rate Data}
\description{
A collection of utilities to simulate data from different compound over-dispersed count distributions representing egg reduction rate (ERR; aka faecal egg count reduction) datasets.
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Authors:
\itemize{
  \item Luc E Coffeng \email{l.coffeng@erasmusmc.nl}
}

}
\keyword{internal}
