% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggSim.R
\name{eggSim}
\alias{eggSim}
\title{Simulate and summarise egg reduction rate data}
\usage{
eggSim(
  reduction,
  budget = 600,
  second_slide_cost = 0.621,
  max_screen = 0.9,
  community_mean = c(24, 48),
  cv_between = c(1.5),
  cv_within = 0.75,
  cv_slide = 0.25,
  cv_reduction = 0,
  count = TRUE,
  log_constant = if (count) 1 else 0,
  screen_threshold = 0,
  grams = 1/24,
  R = 10^3,
  type = "gamma",
  parallelise = TRUE
)
}
\arguments{
\item{reduction}{The arithmetic mean reduction (may be vectorised)}

\item{community_mean}{A vector of arithmetic mean pre-treatment EPG in each community}

\item{cv_between}{A vector of CV reflecting variation in EPG between individuals in each community}

\item{cv_within}{Day-to-day variation in EPG within an individual}

\item{cv_slide}{Variation between faecal samples from the same individual and day}

\item{cv_reduction}{Variation in efficacy between individuals}

\item{count}{Logical flag to base means on count data or the underlying rates}

\item{log_constant}{A constant to add to the count data before calculating geometric means (ignored if count==FALSE)}

\item{screen_threshold}{The threshold count on which to screen individuals}

\item{R}{The number of replicate datasets}

\item{edt}{The egg detection threshold (24 EPG is standard for Kato-Katz)}

\item{N}{A named vector of number of included individuals depending on design type}

\item{maxN}{The maximum number of simulated individuals (on which subsampling occurs)}
}
\value{
A data frame containing the simulated data
}
\description{
Simulate and summarise egg reduction rate data
}
\examples{
means <- eggSim(c(0.8,0.9,0.95), cv_reduction=0)

}
