% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERRparameters.R
\name{ERRparameters}
\alias{ERRparameters}
\title{Calculate simulation parameters for a given set of (artithmetic) mean counts, coefficients of variation and correlation}
\usage{
ERRparameters(mu_pre, mu_post, cv_pre, cv_post, correlation,
  geomgamma = FALSE)
}
\arguments{
\item{mu_pre}{The pre-treatment arithmetic mean count}

\item{mu_post}{The post-treatment arithmetic mean count}

\item{cv_pre}{The extra-Poisson coefficient of variation (ratio of standard deviation to mean) in the pre-treatment data}

\item{cv_post}{The extra-Poisson coefficient of variation (ratio of standard deviation to mean) in the post-treatment data}

\item{correlation}{The extra-Poisson correlation between pre- and post-treatment lambda parameters (GP) or log-lambda parameters (LP)}

\item{geomgamma}{Should the geometric mean reduction for the gamma-Poisson data be estimated by simulation?}
}
\value{
a list with components reflecting the necessary simulation parameters for correlated gamma-Poisson and correlatedlognormal-Poisson distributions
}
\description{
Calculate simulation parameters for a given set of (artithmetic) mean counts, coefficients of variation and correlation
}
\examples{
ERRparameters(100, 10, 2, 2.5, 0.4, TRUE)

}
