% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpDataSim.R
\name{gpDataSim}
\alias{gpDataSim}
\title{Simulate ERR data from a correlated gamma distribution}
\usage{
gpDataSim(N, mu_pre, mu_post, cv_pre, cv_post, correlation,
  parameters = NULL)
}
\arguments{
\item{N}{The number of (paired) observations to simulate}

\item{mu_pre}{The pre-treatment arithmetic mean eggs per gram}

\item{mu_post}{The post-treatment arithmetic mean eggs per gram}

\item{cv_pre}{The extra-Poisson coefficient of variation (ratio of standard deviation to mean) in the pre-treatment data}

\item{cv_post}{The extra-Poisson coefficient of variation (ratio of standard deviation to mean) in the post-treatment data}

\item{correlation}{The extra-Poisson correlation between pre- and post-treatment lambda parameters}

\item{parameters}{Alternative way to specify the parameters as output of the ERRparameters function}
}
\value{
A data frame containing the simulated data
}
\description{
Simulate ERR data from a correlated gamma distribution
}
\examples{
gpDataSim(10, 10, 1, 1, 1, 0.4)

}
