% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpDataSim.R
\name{lpDataSim}
\alias{lpDataSim}
\title{Simulate ERR data from a correlated lognormal distribution}
\usage{
lpDataSim(N, mu_pre, mu_post, cv_pre, cv_post, correlation, parameters = NULL)
}
\arguments{
\item{N}{The number of (paired) observations to simulate}

\item{mu_pre}{The pre-treatment arithmetic mean eggs per gram}

\item{mu_post}{The post-treatment arithmetic mean eggs per gram}

\item{cv_pre}{The extra-Poisson coefficient of variation (ratio of standard deviation to mean) in the pre-treatment data}

\item{cv_post}{The extra-Poisson coefficient of variation (ratio of standard deviation to mean) in the post-treatment data}

\item{correlation}{The extra-Poisson correlation between pre- and post-treatment log-lambda parameters}

\item{parameters}{Alternative way to specify the parameters as output of the ERRparameters function}
}
\value{
A data frame containing the simulated data
}
\description{
Simulate ERR data from a correlated lognormal distribution
}
\examples{
lpDataSim(10, 10, 1, 1, 1, 0.4)

}
