% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_ns.R
\name{survey_ns}
\alias{survey_ns}
\title{Title}
\usage{
survey_ns(
  N_individ = 10,
  N_day_pre = 1,
  N_sample_pre = 1,
  N_day_post = 1,
  N_sample_post = 1,
  mu_pre = 10,
  individ_k = 1,
  day_k = 1,
  sample_k = 1,
  efficacy_a = 1,
  efficacy_b = 1
)
}
\arguments{
\item{efficacy_b}{}
}
\description{
Title
}
