% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_results_levels.R
\name{survey_results_levels}
\alias{survey_results_levels}
\title{Explanation of the meaning of the results from survey_sim}
\usage{
survey_results_levels()
}
\description{
Explanation of the meaning of the results from survey_sim
}
