% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_sim.R
\name{survey_sim}
\alias{survey_sim}
\title{Simulate one or more survey design}
\usage{
survey_sim(
  design = c("NS_11", "SS_11", "SSR_11"),
  parasite = "hookworm",
  method = "kk",
  n_individ = seq(100, 1000, by = 10),
  scenario = survey_scenario(parasite),
  parameters = survey_parameters(design, parasite, method),
  iterations = 1000,
  cl = NULL,
  output = "summarised",
  analysis = "mean",
  check_memory = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{design}{survey design(s) to use.  Ignored if parameters and scenario are specified.}

\item{parasite}{parasite(s) to use.  Ignored if parameters and scenario are specified.}

\item{method}{method(s) to use.  Ignored if parameters and scenario are specified.}

\item{n_individ}{number of individuals (can be a vector)}

\item{scenario}{a data frame of scenarios (see \code{\link{survey_scenario}})}

\item{parameters}{a list of parameter sets (see \code{\link{survey_parameters}})}

\item{iterations}{number of iterations per simulation}

\item{cl}{option to specify either a number of cores or an existing cluster for parallel computation (passed to \code{\link[pbapply]{pblapply}})}

\item{output}{type of output:  one of summarised, full or extended}

\item{analysis}{type of ERR/FECRT analysis:  one of mean or delta}

\item{check_memory}{should the expected memory allocation be checked before proceeding?  Does not apply to output type summarised.}

\item{quiet}{option to suppress output to screen}
}
\description{
Simulate one or more survey design
}
\examples{
results <- survey_sim(design="NS_11", n_individ = c(100,200,300,400,500))

}
