% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgpDataSim.R
\name{cgpDataSim}
\alias{cgpDataSim}
\title{Simulate ERR data from a compound gamma Poisson distribution}
\usage{
cgpDataSim(
  R,
  N,
  reduction,
  community_mean,
  cv_between,
  cv_within,
  cv_slide,
  cv_reduction,
  true_prevalence = 1,
  overall_mean = mean(community_mean),
  grams = 1/24
)
}
\arguments{
\item{R}{The number of replicate datasets}

\item{N}{The maximum number of individuals in total}

\item{reduction}{The arithmetic mean reduction (may be vectorised)}

\item{community_mean}{A vector of arithmetic mean pre-treatment EPG in each community}

\item{cv_between}{A vector of CV reflecting variation in EPG between individuals in each community}

\item{cv_within}{Day-to-day variation in EPG within an individual}

\item{cv_slide}{Variation between faecal samples from the same individual and day}

\item{cv_reduction}{Variation in efficacy between individuals}

\item{true_prevalence}{The true prevalence of infected individuals, i.e. one minus the zero inflation at individual level}

\item{overall_mean}{The overall mean (i.e. the mean of the distribution reflecting community means)}

\item{grams}{The grams of faeces examined (0.0417g i.e. 1/24 EPG is standard for Kato-Katz)}
}
\value{
A data frame containing the simulated data
}
\description{
Simulate ERR data from a compound gamma Poisson distribution
}
\examples{
data <- cgpDataSim(10^3, 100, 0.1, 100, 1, 1, 1, 0, true_prevalence=0.8)

}
