% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_k.R
\name{combined_k}
\alias{combined_k}
\title{Calculate the effective overall k based on coefficients of variation}
\usage{
combined_k(cv_between, cv_within, cv_slide, cv_reduction)
}
\arguments{
\item{cv_between}{A vector of CV reflecting variation in EPG between individuals in each community}

\item{cv_within}{Day-to-day variation in EPG within an individual}

\item{cv_slide}{Variation between faecal samples from the same individual and day}

\item{cv_reduction}{Variation in efficacy between individuals}
}
\value{
a vector of pre_k and post_k
}
\description{
Calculate the effective overall k based on coefficients of variation
}
\examples{
combined_k(1.5, 0.75, 0.25, 0)
combined_k(1, 0.75, 0, 0)
(1/1^2 * 1/0.75^2) / (1/1^2 + 1/0.75^2 + 1)

}
