% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_means.R
\name{design_means}
\alias{design_means}
\title{Calculate the mean egg counts/rates according to various design types}
\usage{
design_means(
  simdata,
  design = c("NS1", "NS2", "NS3", "SS", "SSR1", "SSR2", "SSR3"),
  budget = 600,
  second_slide_cost = 0.621,
  max_screen = 0.9,
  count = TRUE,
  log_constant = if (count) 1 else 0,
  screen_threshold = 0
)
}
\arguments{
\item{simdata}{The data as simulated by cgpDataSim or clpDataSim}

\item{design}{The survey design(s) to be examined}

\item{budget}{The total budget to use for each design}

\item{second_slide_cost}{The cost of a second examination (e.g. Kato-Katz slide) from the same faecal sample, relative to the cost of an entirely independent sample}

\item{max_screen}{The maximum proportion of the budget to use on screening}

\item{count}{Logical flag to base means on count data or the underlying rates}

\item{log_constant}{A constant to add to the count data before calculating geometric means (ignored if count==FALSE)}

\item{screen_threshold}{The threshold count on which to screen individuals}
}
\value{
A data frame of summary statistics
}
\description{
Calculate the mean egg counts/rates according to various design types
}
\examples{
data <- cgpDataSim(10^3, 600, 0.1, 100, 1, 1, 1, 0, true_prevalence=0.8)
means <- design_means(data, second_slide_cost = c(0.1, 0.621, 1))
library("dplyr")
means \%>\%
  group_by(Design, SecondSlideCost) \%>\%
  summarise(N = mean(N), Budget = mean(ScreenBudget + SampleBudget),
    Bias = median(ArithmeticEfficacy - TrueArithmetic),
    SD = sd(ArithmeticEfficacy - TrueArithmetic))

}
