% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_means.R
\name{design_means}
\alias{design_means}
\title{Calculate the mean egg counts/rates according to various design types}
\usage{
design_means(simdata, N = c(NS = 600, SS = 109, SSR1 = 100, SSR2 = 92,
  SSR3 = 95), count = TRUE, log_constant = if (count) 1 else 0,
  screen_threshold = 0)
}
\arguments{
\item{simdata}{The data as simulated by cgpDataSim or clpDataSim}

\item{N}{A named vector of number of included individuals depending on design type}

\item{count}{Logical flag to base means on count data or the underlying rates}

\item{log_constant}{A constant to add to the count data before calculating geometric means (ignored if count==FALSE)}

\item{screen_threshold}{The threshold count on which to screen individuals}
}
\value{
A data frame of summary statistics
}
\description{
Calculate the mean egg counts/rates according to various design types
}
\examples{

}
