% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggSim.R
\name{eggSim}
\alias{eggSim}
\title{Simulate and summarise egg reduction rate data}
\usage{
eggSim(reduction, N = c(NS = 600, SS = 109, SSR1 = 100, SSR2 = 92, SSR3 =
  95), community_mean = c(800, 1000, 1200), cv_between = c(0.8, 1,
  1.2), cv_within = 1, cv_slide = 1, cv_reduction = 1,
  count = TRUE, log_constant = if (count) 1 else 0,
  screen_threshold = 0, edt = 24, R = 10^3, maxN = 10^3)
}
\arguments{
\item{reduction}{The arithmetic mean reduction (may be vectorised)}

\item{N}{The maximum number of individuals in total}

\item{community_mean}{A vector of arithmetic mean pre-treatment EPG in each community}

\item{cv_between}{A vector of CV reflecting variation in EPG between individuals in each community}

\item{cv_within}{Day-to-day variation in EPG within an individual}

\item{cv_slide}{Variation between faecal samples from the same individual and day}

\item{cv_reduction}{Variation in efficacy between individuals}

\item{count}{Logical flag to base means on count data or the underlying rates}

\item{log_constant}{A constant to add to the count data before calculating geometric means (ignored if count==FALSE)}

\item{screen_threshold}{The threshold count on which to screen individuals}

\item{edt}{The egg detection threshold (24 EPG is standard for Kato-Katz)}

\item{R}{The number of replicate datasets}

\item{maxN}{The maximum number of simulated individuals (on which subsampling occurs)}

\item{N}{A named vector of number of included individuals depending on design type}
}
\value{
A data frame containing the simulated data
}
\description{
Simulate and summarise egg reduction rate data
}
\examples{


}
