% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggSim.R
\name{eggSim}
\alias{eggSim}
\title{Simulate and summarise egg reduction rate data}
\usage{
eggSim(
  reduction,
  budget = 600,
  second_slide_cost = 0.621,
  max_screen = 0.9,
  community_mean = c(24, 48),
  cv_between = c(1.5),
  cv_within = 0.75,
  cv_slide = 0.25,
  cv_reduction = 0,
  true_prevalence = 1,
  count = TRUE,
  log_constant = if (count) 1 else 0,
  screen_threshold = 0,
  grams = 1/24,
  R = 10^3,
  design = c("NS", "NS2", "NS3", "SS", "SSR1", "SSR2", "SSR3"),
  summarise = TRUE,
  type = "gamma",
  parallelise = TRUE,
  max_vec = 5e+06
)
}
\arguments{
\item{reduction}{The arithmetic mean reduction (may be vectorised)}

\item{budget}{The total budget to use for each design}

\item{second_slide_cost}{The cost of a second examination (e.g. Kato-Katz slide) from the same faecal sample, relative to the cost of an entirely independent sample}

\item{max_screen}{The maximum proportion of the budget to use on screening}

\item{community_mean}{A vector of arithmetic mean pre-treatment EPG in each community}

\item{cv_between}{A vector of CV reflecting variation in EPG between individuals in each community}

\item{cv_within}{Day-to-day variation in EPG within an individual}

\item{cv_slide}{Variation between faecal samples from the same individual and day}

\item{cv_reduction}{Variation in efficacy between individuals}

\item{true_prevalence}{The true prevalence of infected individuals, i.e. one minus the zero inflation at individual level}

\item{count}{Logical flag to base means on count data or the underlying rates}

\item{log_constant}{A constant to add to the count data before calculating geometric means (ignored if count==FALSE)}

\item{screen_threshold}{The threshold count on which to screen individuals}

\item{grams}{The volume of faeces actually examined for eggs (1/24 grams is standard for Kato-Katz)}

\item{R}{The number of replicate datasets}

\item{design}{The survey design(s) to be examined}

\item{summarise}{Should the simulation results be summarised?}

\item{type}{One of gamma or lognormal - at present only the former is suppoered}

\item{parallelise}{Option to parallelise simulations (can also be an integer giving the number of cores to use, or a cluster pre-created using the parallal package)}

\item{max_vec}{The maximum number of data frame rows to use for vectorisation (affects memory usage)}
}
\value{
A data frame containing the simulated data
}
\description{
Simulate and summarise egg reduction rate data
}
\examples{
means <- eggSim(c(0.2,0.1,0.05), cv_reduction=0, R=10^2, parallelise=FALSE)

}
