% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_parameters.R
\name{survey_parameters}
\alias{survey_parameters}
\alias{survey_scenario}
\title{Parameter sets for survey simulations}
\usage{
survey_parameters(
  design = c("SS_11", "SS_12", "NS_11", "NS_12", "SSR_11", "SSR_12"),
  parasite = c("ascaris", "trichuris", "hookworm"),
  method = c("kk", "miniflotac", "fecpak")
)

survey_scenario(parasite = c("ascaris", "trichuris", "hookworm"))
}
\arguments{
\item{n_aliquot_post}{}
}
\description{
Parameter sets for survey simulations
}
