% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_parameters.R
\name{survey_parameters}
\alias{survey_parameters}
\alias{survey_scenario}
\title{Parameter sets for survey simulations}
\usage{
survey_parameters(
  design = c("SS_A", "NS_A", "NS_B", "SSR_A", "SSR_B"),
  parasite = c("ascaris", "trichuris", "hookworm"),
  method = c("kk", "miniflotac", "fecpak")
)

survey_scenario(parasite)
}
\arguments{
\item{n_aliquot_post}{}
}
\description{
Parameter sets for survey simulations
}
