% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_sim.R
\name{survey_sim}
\alias{survey_sim}
\title{Title}
\usage{
survey_sim(
  design = "NS_2x2",
  parasite = "HW",
  method = "KK",
  iterations = 1000,
  n_individ = 100,
  mu_pre = 100,
  reduction = 0.9,
  params_design = parameters_design(design),
  params_parasite = parameters_parasite(parasite),
  params_method = parameters_method(method),
  cost = survey_cost(method),
  pb = NA,
  parameter_output = NA
)
}
\arguments{
\item{parameter_output}{}

\item{family}{currently ignored}
}
\description{
Title
}
\examples{
results <- survey_sim(design = c("NS_1x1", "NS_2x2"))

}
