% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_sim.R
\name{survey_sim}
\alias{survey_sim}
\title{Title}
\usage{
survey_sim(
  design = c("NS_A", "SS_A", "SSR_A"),
  parasite = "hookworm",
  method = "kk",
  n_individ = 100,
  scenario = survey_scenario(parasite),
  parameters = survey_parameters(design, parasite, method),
  iterations = 1000,
  pb = NA,
  output = NA
)
}
\arguments{
\item{family}{currently ignored}

\item{parameter_output}{}
}
\description{
Title
}
\examples{
results <- survey_sim(design = c("NS_1x1", "NS_2x2"))

}
