% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_sim.R
\name{survey_sim}
\alias{survey_sim}
\title{Title}
\usage{
survey_sim(
  design = "NS",
  iterations = 10,
  n_individ = 10,
  n_day_pre = 1,
  n_aliquot_pre = 1,
  n_day_post = 1,
  n_aliquot_post = 1,
  mu_pre = 10,
  weight = 1,
  performance = 1,
  cost_sample = 1,
  cost_aliquot = 1,
  individ_k = 1,
  day_k = 1,
  aliquot_k = 1,
  efficacy_a = 1,
  efficacy_b = 1,
  pb = length(design) > 1,
  parameter_output = FALSE
)
}
\arguments{
\item{parameter_output}{}
}
\description{
Title
}
